+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Additional data from Greece (contributor: K. J. Westner)
 * Additional data from Bulgaria (contributor: K. J. Westner)
+* The detection pattern of the function `LI_ratios_all()` was improved. 
+
 
 # Version 1.0